properly handles plugins
at some point the previous night, in error, one of two functions was committed to handle zips, and unfortunately the non working copy got pushed. tonight we recreated the functioning zip and made it even better to properly handle some people who just put the plugin files in the zip and some people who put the plugin files in a folder. this handles that. unless of course we've deleted the wrong code again.