fix open file limit exceeded
For blogs with more than 512 pages, the default open file limit for most
Kernels (1024) is exceeded. In this case the operating system just
refuses to open any more files.

Instead of using defer and wait for the files to be closed, run the
(*os.File).Close function after completing working with the file instead
of waiting for the function to return before closing the files.

Fixes: https://github.com/donuts-are-good/bearclaw/issues/11
Fixes: b60fc70 ("hello world")
Signed-off-by: Moritz Poldrack <git@moritz.sh>